requirements.txt
pygame-ce>=2.5.0
sounddevice>=0.5.0
numpy>=2.0
mido>=1.3.0
